# perf: replace lookahead by lookaheadCharCode