--- a/packages/babel-parser/src/tokenizer/index.js
+++ b/packages/babel-parser/src/tokenizer/index.js
@@ -17,6 +17,7 @@ import {
 import State from "./state";
 
 const VALID_REGEX_FLAGS = new Set(["g", "m", "s", "i", "y", "u"]);
+const skipWhiteSpace = /(?:\s|\/\/.*|\/\*[^]*?\*\/)*/g;
 
 // The following character codes are forbidden from being
 // an immediate sibling of NumericLiteralSeparator _
@@ -166,6 +167,14 @@ export default class Tokenizer extends LocationParser {
     const curr = this.state;
     this.state = old;
     return curr;
+  }
+
+  lookaheadCharCode(): number {
+    const thisTokEnd = this.state.pos;
+    skipWhiteSpace.lastIndex = thisTokEnd;
+    const skip = skipWhiteSpace.exec(this.input);
+    const next = thisTokEnd + skip[0].length;
+    return this.input.charCodeAt(next);
   }
 
   // Toggle strict mode. Re-reads the next number or string to please