--- a/packages/babel-parser/src/tokenizer/index.js
+++ b/packages/babel-parser/src/tokenizer/index.js
@@ -173,6 +173,7 @@ export default class Tokenizer extends LocationParser {
     const thisTokEnd = this.state.pos;
     skipWhiteSpace.lastIndex = thisTokEnd;
     const skip = skipWhiteSpace.exec(this.input);
+    // $FlowIgnore: The skipWhiteSpace ensures to match any string
     const next = thisTokEnd + skip[0].length;
     return this.input.charCodeAt(next);
   }