<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/master/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | 
| Patch: Bug Fix?          | Perf
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | No
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

Introduced `tokenizer.lookaheadCharCode()` method, which simply skips the whitespaces and comments, until the next non-whitespace charcode. The idea is [borrowed from acorn](https://github.com/acornjs/acorn/blob/7e9817d17639d95cc6dbacfde734a0626b2a7dea/acorn/src/statement.js#L38-L40).

`tokenizer.lookahead()` is expensive in the way that it would cache the whole state and restore after `nextToken`. However, in most of times `tokenizer.lookahead` is used to compare the `type` of next token only, therefore we can replace `tokenizer.lookahead()` by `tokenizer.lookaheadCharCode()` which does _not_ interfere with the tokenizer state.

## Logic difference

The replacement is logically equivalent when the character maps to only one possible token type.

It is okay to replace `tt.parenL` with `charcode === 40`.

It will introduce logic difference when replacing `tt.dot` with `charcode === 46` since it may be part of `tt.ellipsis`. However `tt.ellipsis` should not immediately follow the `function` keyword and would bail in `parseExpression`. The same applies when following `import` and `null`.

## Benchmark report
Environment
CPU: Intel Core-i7 6820HQ
RAM: 16 GB 2133 MHz LPDDR3
Kernel: Darwin 18.7.0
Node.js: v12.9.0

Control group versions
acorn: v7.0.0
babel-parser: v7.5.5
```
┌──────────────────────────┬─────────────────────────────┬──────────────────────────────┬──────────────────────────────┬──────────────────────────────┬──────────────────────────────┐
│ fixture                  │ acorn                       │ babel                        │ babelDev                     │ esprima                      │ meriyah                      │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/angular.js           │ 30.26 ops/sec ±4.75% (33ms) │ 21.27 ops/sec ±4.57% (47ms)  │ 23.56 ops/sec ±3.79% (42ms)  │ 24.56 ops/sec ±7.19% (41ms)  │ 41.04 ops/sec ±2.64% (24ms)  │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/ember.debug.js       │ 11.48 ops/sec ±3.63% (87ms) │ 8.04 ops/sec ±3.97% (124ms)  │ 9.01 ops/sec ±3.5% (111ms)   │ 8.87 ops/sec ±14.86% (113ms) │ 17.27 ops/sec ±8.29% (58ms)  │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/babylon-dist.js      │ 54.39 ops/sec ±3.17% (18ms) │ 33.18 ops/sec ±5.38% (30ms)  │ 34.1 ops/sec ±5.29% (29ms)   │ 44.85 ops/sec ±1.74% (22ms)  │ 93.23 ops/sec ±1.13% (11ms)  │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/jquery.js            │ 65.05 ops/sec ±2.82% (15ms) │ 42.83 ops/sec ±5.56% (23ms)  │ 46.42 ops/sec ±4.71% (22ms)  │ 57.87 ops/sec ±1.11% (17ms)  │ 115 ops/sec ±2% (8.683ms)    │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/backbone.js          │ 270 ops/sec ±2.2% (3.707ms) │ 192 ops/sec ±2.09% (5.197ms) │ 203 ops/sec ±2.61% (4.915ms) │ 227 ops/sec ±2.78% (4.402ms) │ 465 ops/sec ±2.11% (2.148ms) │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/react-with-addons.js │ 32.19 ops/sec ±4.13% (31ms) │ 20.24 ops/sec ±4.75% (49ms)  │ 22.79 ops/sec ±3.07% (44ms)  │ 25.02 ops/sec ±4.86% (40ms)  │ 52.12 ops/sec ±1.53% (19ms)  │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es6/angular-compiler.js  │ 14.17 ops/sec ±5.99% (71ms) │ 10.35 ops/sec ±4.55% (97ms)  │ 11.39 ops/sec ±5.81% (88ms)  │ 14.55 ops/sec ±3.34% (69ms)  │ 26.89 ops/sec ±2.4% (37ms)   │
├──────────────────────────┼─────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es6/material-ui-core.js  │ 21.06 ops/sec ±3.95% (47ms) │ 13.44 ops/sec ±3% (74ms)     │ 14.73 ops/sec ±2.93% (68ms)  │ 20.12 ops/sec ±3.46% (50ms)  │ 34.92 ops/sec ±2.13% (29ms)  │
└──────────────────────────┴─────────────────────────────┴──────────────────────────────┴──────────────────────────────┴──────────────────────────────┴──────────────────────────────┘
```